Migrate app to typescript
