+++ react/package.json
@@ -8,6 +8,9 @@
   "author": "VTEX",
   "license": "ISC",
   "devDependencies": {
+    "@types/prop-types": "^15.5.6",
+    "@types/react": "^16.4.14",
+    "@types/react-helmet": "^5.0.7",
     "babel-jest": "^22.4.3",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react-app": "^3.1.1",
@@ -16,7 +19,8 @@
     "react": "^16.3.1",
     "react-dom": "^16.3.1",
     "react-intl": "^2.4.0",
-    "react-testing-library": "^2.1.1"
+    "react-testing-library": "^2.1.1",
+    "tslint-eslint-rules": "^5.4.0"
   },
   "jest": {
     "verbose": true,
